requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
pytesseract
